:art:(governor): improve key extraction in rate limiting
The CookieExtractor is updated to first try to extract the key from the
'sid' cookie, and if that fails, it falls back to using the client's IP
address.

For client IP address there are fallback procedures: if we don't get
x-forwarded-for etc. headers (but we should from cloudflare), we fall
back to more restrictive peer ip

Closes #41